requirements.txt
paho-mqtt==1.5.1
pynput==1.7.4
